chore: moved resources and proxies around
