+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/dataservice/ConceptResource.java
@@ -1,10 +1,10 @@
-package software.uncharted.terarium.hmiserver.resources;
+package software.uncharted.terarium.hmiserver.resources.dataservice;
 
 import io.quarkus.security.Authenticated;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import software.uncharted.terarium.hmiserver.models.dataservice.Concept;
-import software.uncharted.terarium.hmiserver.proxies.ConceptProxy;
+import software.uncharted.terarium.hmiserver.proxies.dataservice.ConceptProxy;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
