+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/dataservice/ExternalResource.java
@@ -1,11 +1,11 @@
-package software.uncharted.terarium.hmiserver.resources;
+package software.uncharted.terarium.hmiserver.resources.dataservice;
 
 import io.quarkus.security.Authenticated;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import software.uncharted.terarium.hmiserver.models.dataservice.Publication;
 import software.uncharted.terarium.hmiserver.models.dataservice.Software;
-import software.uncharted.terarium.hmiserver.proxies.ExternalProxy;
+import software.uncharted.terarium.hmiserver.proxies.dataservice.ExternalProxy;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
