+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/dataservice/ProjectResource.java
@@ -1,11 +1,11 @@
-package software.uncharted.terarium.hmiserver.resources;
+package software.uncharted.terarium.hmiserver.resources.dataservice;
 
 import io.quarkus.security.Authenticated;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 import software.uncharted.terarium.hmiserver.models.dataservice.Project;
 import software.uncharted.terarium.hmiserver.models.dataservice.ResourceType;
-import software.uncharted.terarium.hmiserver.proxies.ProjectProxy;
+import software.uncharted.terarium.hmiserver.proxies.dataservice.ProjectProxy;
 
 import javax.inject.Inject;
 import javax.ws.rs.*;
