+++ packages/services/hmi-server/src/main/java/software/uncharted/terarium/hmiserver/resources/dataservice/ProvenanceResource.java
@@ -1,9 +1,9 @@
-package software.uncharted.terarium.hmiserver.resources;
+package software.uncharted.terarium.hmiserver.resources.dataservice;
 
 import io.quarkus.security.Authenticated;
 import org.eclipse.microprofile.openapi.annotations.tags.Tag;
 import org.eclipse.microprofile.rest.client.inject.RestClient;
-import software.uncharted.terarium.hmiserver.proxies.ProvenanceProxy;
+import software.uncharted.terarium.hmiserver.proxies.dataservice.ProvenanceProxy;
 
 import javax.inject.Inject;
 import javax.ws.rs.GET;
